feat: criar o componente CollectionCreateModal para reaproveitamento de código
